Divided class structure into 2 separate classes  Translator handles all the translation stuff and TranslateClient handles the Translator static and non-static calls
